pyproject.toml
[tool.poetry]
name = "iacraft"
version = "0.1.0"
description = ""
authors = ["Peter Chung <touhonoob@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain = "^0.0.115"
click = "^8.1.3"
chromadb = "^0.3.11"
openai = {extras = ["embeddings"], version = "^0.27.2"}
httpx = "^0.23.3"
aiometer = "^0.4.0"
markdown-it-py = "^2.2.0"
diff-match-patch = "^20200713"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
